requirements.txt
astroid
httplib2
isort
lazy-object-proxy
mccabe
oauth2client
pyasn1
pyasn1-modules
pylint
python-dateutil
rsa
simplejson
uritemplate
wrapt

# This is actually needed. The rest I'm not sure...
pyyaml
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2